requirements.txt
aniso8601==10.0.1
attrs==25.4.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
flask-restx==1.3.2
fonttools==4.60.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
rpds-py==0.28.0
scipy==1.16.3
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
watchdog==6.0.0
Werkzeug==3.1.3
